Move cfpq_request function to another module
